auto-update of row/col metadata, cleaning
Now front-end updates are automatically synced to the row/col
metadata (no need to run net.get_manual_category()). We changed
the name from custom_category to manual_category for consistenty.
We are working on syncing the categories between two widgets,
which requires front-end to back-end to front-end changes. We
also renamed ExampleWidget CGM2 which can be used as follows:
net1 = Network(CGM2)